+++ nomad/state/state_store.go
@@ -14,7 +14,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/hashicorp/nomad/helper"
-	"github.com/hashicorp/nomad/nomad/event"
+	"github.com/hashicorp/nomad/nomad/stream"
 	"github.com/hashicorp/nomad/nomad/structs"
 )
 
@@ -46,6 +46,16 @@ type StateStoreConfig struct {
 
 	// Region is the region of the server embedding the state store.
 	Region string
+
+	// EnablePublisher is used to enable or disable the event publisher
+	EnablePublisher bool
+
+	// EventBufferSize configures the amount of events to hold in memory
+	EventBufferSize int64
+
+	// DurableEventCount is used to determine if events from transaction changes
+	// should be saved in go-memdb
+	DurableEventCount int64
 }
 
 // The StateStore is responsible for maintaining all the Nomad
@@ -65,6 +75,10 @@ type StateStore struct {
 	// abandonCh is used to signal watchers that this state store has been
 	// abandoned (usually during a restore). This is only ever closed.
 	abandonCh chan struct{}
+
+	// TODO: refactor abondonCh to use a context so that both can use the same
+	// cancel mechanism.
+	stopEventBroker func()
 }
 
 // NewStateStore is used to create a new state store
@@ -76,12 +90,25 @@ func NewStateStore(config *StateStoreConfig) (*StateStore, error) {
 	}
 
 	// Create the state store
+	ctx, cancel := context.WithCancel(context.TODO())
 	s := &StateStore{
-		logger:    config.Logger.Named("state_store"),
-		config:    config,
-		abandonCh: make(chan struct{}),
+		logger:          config.Logger.Named("state_store"),
+		config:          config,
+		abandonCh:       make(chan struct{}),
+		stopEventBroker: cancel,
 	}
-	s.db = NewChangeTrackerDB(db, event.NewPublisher(), processDBChanges)
+
+	if config.EnablePublisher {
+		// Create new event publisher using provided config
+		broker := stream.NewEventBroker(ctx, stream.EventBrokerCfg{
+			EventBufferSize: config.EventBufferSize,
+			Logger:          config.Logger,
+			OnEvict:         s.eventBrokerEvict,
+		})
+		s.db = NewChangeTrackerDB(db, broker, processDBChanges, config.DurableEventCount)
+	} else {
+		s.db = NewChangeTrackerDB(db, nil, noOpProcessChanges, 0)
+	}
 
 	// Initialize the state store with required enterprise objects
 	if err := s.enterpriseInit(); err != nil {
@@ -89,6 +116,37 @@ func NewStateStore(config *StateStoreConfig) (*StateStore, error) {
 	}
 
 	return s, nil
+}
+
+func (s *StateStore) EventBroker() (*stream.EventBroker, error) {
+	if s.db.publisher == nil {
+		return nil, fmt.Errorf("EventBroker not configured")
+	}
+	return s.db.publisher, nil
+}
+
+// eventBrokerEvict is used as a callback to delete an evicted events
+// entry from go-memdb.
+func (s *StateStore) eventBrokerEvict(events *structs.Events) {
+	if err := s.deleteEvent(events); err != nil {
+		if err == memdb.ErrNotFound {
+			s.logger.Info("Evicted event was not found in go-memdb table", "event index", events.Index)
+		} else {
+			s.logger.Error("Error deleting event from events table", "error", err)
+		}
+	}
+}
+
+func (s *StateStore) deleteEvent(events *structs.Events) error {
+	txn := s.db.memdb.Txn(true)
+	defer txn.Abort()
+
+	if err := txn.Delete("events", events); err != nil {
+		return err
+	}
+
+	txn.Commit()
+	return nil
 }
 
 // Config returns the state store configuration.
@@ -100,14 +158,15 @@ func (s *StateStore) Config() *StateStoreConfig {
 // we use MemDB, we just need to snapshot the state of the underlying
 // database.
 func (s *StateStore) Snapshot() (*StateSnapshot, error) {
-	memDBSnap := s.db.db.Snapshot()
+	memDBSnap := s.db.memdb.Snapshot()
 
 	store := StateStore{
 		logger: s.logger,
 		config: s.config,
 	}
 
-	store.db = NewChangeTrackerDB(memDBSnap, &noOpPublisher{}, noOpProcessChanges)
+	// Create a new change tracker DB that does not publish or track changes
+	store.db = NewChangeTrackerDB(memDBSnap, nil, noOpProcessChanges, 0)
 
 	snap := &StateSnapshot{
 		StateStore: store,
@@ -190,7 +249,14 @@ func (s *StateStore) AbandonCh() <-chan struct{} {
 // Abandon is used to signal that the given state store has been abandoned.
 // Calling this more than one time will panic.
 func (s *StateStore) Abandon() {
+	s.StopEventBroker()
 	close(s.abandonCh)
+}
+
+// StopStopEventBroker calls the cancel func for the state stores event
+// publisher. It should be called during server shutdown.
+func (s *StateStore) StopEventBroker() {
+	s.stopEventBroker()
 }
 
 // QueryFn is the definition of a function that can be used to implement a basic
@@ -238,7 +304,7 @@ RUN_QUERY:
 }
 
 // UpsertPlanResults is used to upsert the results of a plan.
-func (s *StateStore) UpsertPlanResults(index uint64, results *structs.ApplyPlanResultsRequest) error {
+func (s *StateStore) UpsertPlanResults(msgType structs.MessageType, index uint64, results *structs.ApplyPlanResultsRequest) error {
 	snapshot, err := s.Snapshot()
 	if err != nil {
 		return err
@@ -260,7 +326,7 @@ func (s *StateStore) UpsertPlanResults(index uint64, results *structs.ApplyPlanR
 		return err
 	}
 
-	txn := s.db.WriteTxn(index)
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	// Upsert the newly created or updated deployment
@@ -324,8 +390,7 @@ func (s *StateStore) UpsertPlanResults(index uint64, results *structs.ApplyPlanR
 		}
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // addComputedAllocAttrs adds the computed/derived attributes to the allocation.
@@ -394,8 +459,7 @@ func (s *StateStore) UpsertJobSummary(index uint64, jobSummary *structs.JobSumma
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeleteJobSummary deletes the job summary with the given ID. This is for
@@ -411,8 +475,7 @@ func (s *StateStore) DeleteJobSummary(index uint64, namespace, id string) error 
 	if err := txn.Insert("index", &IndexEntry{"job_summary", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // UpsertDeployment is used to insert a new deployment. If cancelPrior is set to
@@ -423,8 +486,7 @@ func (s *StateStore) UpsertDeployment(index uint64, deployment *structs.Deployme
 	if err := s.upsertDeploymentImpl(index, deployment, txn); err != nil {
 		return err
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 func (s *StateStore) upsertDeploymentImpl(index uint64, deployment *structs.Deployment, txn *txn) error {
@@ -636,8 +698,7 @@ func (s *StateStore) DeleteDeployment(index uint64, deploymentIDs []string) erro
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // UpsertScalingEvent is used to insert a new scaling event.
@@ -688,8 +749,7 @@ func (s *StateStore) UpsertScalingEvent(index uint64, req *structs.ScalingEventR
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // ScalingEvents returns an iterator over all the job scaling events
@@ -723,6 +783,21 @@ func (s *StateStore) ScalingEventsByJob(ws memdb.WatchSet, namespace, jobID stri
 	return nil, 0, nil
 }
 
+// UpsertNodeMsgType is used to register a node or update a node definition
+// This is assumed to be triggered by the client, so we retain the value
+// of drain/eligibility which is set by the scheduler.
+// TODO(drew) remove this and update all test callers of UpsertNode to use msgType
+func (s *StateStore) UpsertNodeMsgType(msgType structs.MessageType, index uint64, node *structs.Node) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+
+	err := upsertNodeTxn(txn, index, node)
+	if err != nil {
+		return nil
+	}
+	return txn.Commit()
+}
+
 // UpsertNode is used to register a node or update a node definition
 // This is assumed to be triggered by the client, so we retain the value
 // of drain/eligibility which is set by the scheduler.
@@ -730,6 +805,14 @@ func (s *StateStore) UpsertNode(index uint64, node *structs.Node) error {
 	txn := s.db.WriteTxn(index)
 	defer txn.Abort()
 
+	err := upsertNodeTxn(txn, index, node)
+	if err != nil {
+		return nil
+	}
+	return txn.Commit()
+}
+
+func upsertNodeTxn(txn *txn, index uint64, node *structs.Node) error {
 	// Check if the node already exists
 	existing, err := txn.First("nodes", "id", node.ID)
 	if err != nil {
@@ -778,18 +861,37 @@ func (s *StateStore) UpsertNode(index uint64, node *structs.Node) error {
 		return fmt.Errorf("csi plugin update failed: %v", err)
 	}
 
-	txn.Commit()
 	return nil
 }
 
 // DeleteNode deregisters a batch of nodes
+func (s *StateStore) DeleteNodeMsgType(msgType structs.MessageType, index uint64, nodes []string) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+
+	err := deleteNodeTxn(txn, index, nodes)
+	if err != nil {
+		return nil
+	}
+	return txn.Commit()
+}
+
+// DeleteNode deregisters a batch of nodes
 func (s *StateStore) DeleteNode(index uint64, nodes []string) error {
+	txn := s.db.WriteTxn(index)
+	defer txn.Abort()
+
+	err := deleteNodeTxn(txn, index, nodes)
+	if err != nil {
+		return nil
+	}
+	return txn.Commit()
+}
+
+func deleteNodeTxn(txn *txn, index uint64, nodes []string) error {
 	if len(nodes) == 0 {
 		return fmt.Errorf("node ids missing")
 	}
-
-	txn := s.db.WriteTxn(index)
-	defer txn.Abort()
 
 	for _, nodeID := range nodes {
 		existing, err := txn.First("nodes", "id", nodeID)
@@ -815,21 +917,19 @@ func (s *StateStore) DeleteNode(index uint64, nodes []string) error {
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
 	return nil
 }
 
 // UpdateNodeStatus is used to update the status of a node
-func (s *StateStore) UpdateNodeStatus(index uint64, nodeID, status string, updatedAt int64, event *structs.NodeEvent) error {
-	txn := s.db.WriteTxn(index)
+func (s *StateStore) UpdateNodeStatus(msgType structs.MessageType, index uint64, nodeID, status string, updatedAt int64, event *structs.NodeEvent) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	if err := s.updateNodeStatusTxn(txn, nodeID, status, updatedAt, event); err != nil {
 		return err
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 func (s *StateStore) updateNodeStatusTxn(txn *txn, nodeID, status string, updatedAt int64, event *structs.NodeEvent) error {
@@ -868,16 +968,27 @@ func (s *StateStore) updateNodeStatusTxn(txn *txn, nodeID, status string, update
 }
 
 // BatchUpdateNodeDrain is used to update the drain of a node set of nodes
-func (s *StateStore) BatchUpdateNodeDrain(index uint64, updatedAt int64, updates map[string]*structs.DrainUpdate, events map[string]*structs.NodeEvent) error {
-	txn := s.db.WriteTxn(index)
+func (s *StateStore) BatchUpdateNodeDrain(msgType structs.MessageType, index uint64, updatedAt int64, updates map[string]*structs.DrainUpdate, events map[string]*structs.NodeEvent) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 	for node, update := range updates {
 		if err := s.updateNodeDrainImpl(txn, index, node, update.DrainStrategy, update.MarkEligible, updatedAt, events[node]); err != nil {
 			return err
 		}
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
+}
+
+// UpdateNodeDrain is used to update the drain of a node
+func (s *StateStore) UpdateNodeDrainMsgType(msgType structs.MessageType, index uint64, nodeID string,
+	drain *structs.DrainStrategy, markEligible bool, updatedAt int64, event *structs.NodeEvent) error {
+
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+	if err := s.updateNodeDrainImpl(txn, index, nodeID, drain, markEligible, updatedAt, event); err != nil {
+		return err
+	}
+	return txn.Commit()
 }
 
 // UpdateNodeDrain is used to update the drain of a node
@@ -889,8 +1000,7 @@ func (s *StateStore) UpdateNodeDrain(index uint64, nodeID string,
 	if err := s.updateNodeDrainImpl(txn, index, nodeID, drain, markEligible, updatedAt, event); err != nil {
 		return err
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 func (s *StateStore) updateNodeDrainImpl(txn *txn, index uint64, nodeID string,
@@ -938,9 +1048,9 @@ func (s *StateStore) updateNodeDrainImpl(txn *txn, index uint64, nodeID string,
 }
 
 // UpdateNodeEligibility is used to update the scheduling eligibility of a node
-func (s *StateStore) UpdateNodeEligibility(index uint64, nodeID string, eligibility string, updatedAt int64, event *structs.NodeEvent) error {
+func (s *StateStore) UpdateNodeEligibility(msgType structs.MessageType, index uint64, nodeID string, eligibility string, updatedAt int64, event *structs.NodeEvent) error {
 
-	txn := s.db.WriteTxn(index)
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	// Lookup the node
@@ -979,8 +1089,20 @@ func (s *StateStore) UpdateNodeEligibility(index uint64, nodeID string, eligibil
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
+}
+
+func (s *StateStore) UpsertNodeEventsMsgType(msgType structs.MessageType, index uint64, nodeEvents map[string][]*structs.NodeEvent) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+
+	for nodeID, events := range nodeEvents {
+		if err := s.upsertNodeEvents(index, nodeID, events, txn); err != nil {
+			return err
+		}
+	}
+
+	return txn.Commit()
 }
 
 // UpsertNodeEvents adds the node events to the nodes, rotating events as
@@ -995,8 +1117,7 @@ func (s *StateStore) UpsertNodeEvents(index uint64, nodeEvents map[string][]*str
 		}
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // upsertNodeEvent upserts a node event for a respective node. It also maintains
@@ -1382,8 +1503,17 @@ func (s *StateStore) UpsertJob(index uint64, job *structs.Job) error {
 	if err := s.upsertJobImpl(index, job, false, txn); err != nil {
 		return err
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
+}
+
+// UpsertJob is used to register a job or update a job definition
+func (s *StateStore) UpsertJobMsgType(msgType structs.MessageType, index uint64, job *structs.Job) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+	if err := s.upsertJobImpl(index, job, false, txn); err != nil {
+		return err
+	}
+	return txn.Commit()
 }
 
 // UpsertJobTxn is used to register a job or update a job definition, like UpsertJob,
@@ -1488,7 +1618,7 @@ func (s *StateStore) DeleteJob(index uint64, namespace, jobID string) error {
 
 	err := s.DeleteJobTxn(index, namespace, jobID, txn)
 	if err == nil {
-		txn.Commit()
+		return txn.Commit()
 	}
 	return err
 }
@@ -2016,8 +2146,7 @@ func (s *StateStore) CSIVolumeRegister(index uint64, volumes []*structs.CSIVolum
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // CSIVolumes returns the unfiltered list of all volumes
@@ -2206,8 +2335,7 @@ func (s *StateStore) CSIVolumeClaim(index uint64, namespace, id string, claim *s
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // CSIVolumeDeregister removes the volume from the server
@@ -2249,8 +2377,7 @@ func (s *StateStore) CSIVolumeDeregister(index uint64, namespace string, ids []s
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // volSafeToForce checks if the any of the remaining allocations
@@ -2431,7 +2558,7 @@ func (s *StateStore) CSIPluginDenormalize(ws memdb.WatchSet, plug *structs.CSIPl
 		if alloc == nil {
 			continue
 		}
-		plug.Allocations = append(plug.Allocations, alloc.Stub())
+		plug.Allocations = append(plug.Allocations, alloc.Stub(nil))
 	}
 
 	return plug, nil
@@ -2461,8 +2588,7 @@ func (s *StateStore) UpsertCSIPlugin(index uint64, plug *structs.CSIPlugin) erro
 	if err := txn.Insert("index", &IndexEntry{"csi_plugins", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeleteCSIPlugin deletes the plugin if it's not in use.
@@ -2492,8 +2618,7 @@ func (s *StateStore) DeleteCSIPlugin(index uint64, id string) error {
 	if err != nil {
 		return fmt.Errorf("csi_plugins delete error: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // UpsertPeriodicLaunch is used to register a launch or update it.
@@ -2524,8 +2649,7 @@ func (s *StateStore) UpsertPeriodicLaunch(index uint64, launch *structs.Periodic
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeletePeriodicLaunch is used to delete the periodic launch
@@ -2535,7 +2659,7 @@ func (s *StateStore) DeletePeriodicLaunch(index uint64, namespace, jobID string)
 
 	err := s.DeletePeriodicLaunchTxn(index, namespace, jobID, txn)
 	if err == nil {
-		txn.Commit()
+		return txn.Commit()
 	}
 	return err
 }
@@ -2603,7 +2727,19 @@ func (s *StateStore) UpsertEvals(index uint64, evals []*structs.Evaluation) erro
 
 	err := s.UpsertEvalsTxn(index, evals, txn)
 	if err == nil {
-		txn.Commit()
+		return txn.Commit()
+	}
+	return err
+}
+
+// UpsertEvals is used to upsert a set of evaluations
+func (s *StateStore) UpsertEvalsMsgType(msgType structs.MessageType, index uint64, evals []*structs.Evaluation) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+
+	err := s.UpsertEvalsTxn(index, evals, txn)
+	if err == nil {
+		return txn.Commit()
 	}
 	return err
 }
@@ -2801,8 +2937,7 @@ func (s *StateStore) DeleteEval(index uint64, evals []string, allocs []string) e
 		return fmt.Errorf("setting job status failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // EvalByID is used to lookup an eval by its ID
@@ -2920,8 +3055,8 @@ func (s *StateStore) EvalsByNamespace(ws memdb.WatchSet, namespace string) (memd
 // most things, some updates are authoritative from the client. Specifically,
 // the desired state comes from the schedulers, while the actual state comes
 // from clients.
-func (s *StateStore) UpdateAllocsFromClient(index uint64, allocs []*structs.Allocation) error {
-	txn := s.db.WriteTxn(index)
+func (s *StateStore) UpdateAllocsFromClient(msgType structs.MessageType, index uint64, allocs []*structs.Allocation) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	// Handle each of the updated allocations
@@ -2936,8 +3071,7 @@ func (s *StateStore) UpdateAllocsFromClient(index uint64, allocs []*structs.Allo
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // nestedUpdateAllocFromClient is used to nest an update of an allocation with client status
@@ -3036,8 +3170,18 @@ func (s *StateStore) UpsertAllocs(index uint64, allocs []*structs.Allocation) er
 	if err := s.upsertAllocsImpl(index, allocs, txn); err != nil {
 		return err
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
+}
+
+// UpsertAllocsMsgType is used to evict a set of allocations and allocate new ones at
+// the same time.
+func (s *StateStore) UpsertAllocsMsgType(msgType structs.MessageType, index uint64, allocs []*structs.Allocation) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
+	defer txn.Abort()
+	if err := s.upsertAllocsImpl(index, allocs, txn); err != nil {
+		return err
+	}
+	return txn.Commit()
 }
 
 // upsertAllocs is the actual implementation of UpsertAllocs so that it may be
@@ -3163,10 +3307,10 @@ func (s *StateStore) upsertAllocsImpl(index uint64, allocs []*structs.Allocation
 
 // UpdateAllocsDesiredTransitions is used to update a set of allocations
 // desired transitions.
-func (s *StateStore) UpdateAllocsDesiredTransitions(index uint64, allocs map[string]*structs.DesiredTransition,
+func (s *StateStore) UpdateAllocsDesiredTransitions(msgType structs.MessageType, index uint64, allocs map[string]*structs.DesiredTransition,
 	evals []*structs.Evaluation) error {
 
-	txn := s.db.WriteTxn(index)
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	// Handle each of the updated allocations
@@ -3187,8 +3331,7 @@ func (s *StateStore) UpdateAllocsDesiredTransitions(index uint64, allocs map[str
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // nestedUpdateAllocDesiredTransition is used to nest an update of an
@@ -3308,6 +3451,10 @@ func (s *StateStore) AllocsByIDPrefixInNSes(ws memdb.WatchSet, namespaces map[st
 func (s *StateStore) AllocsByNode(ws memdb.WatchSet, node string) ([]*structs.Allocation, error) {
 	txn := s.db.ReadTxn()
 
+	return allocsByNodeTxn(txn, ws, node)
+}
+
+func allocsByNodeTxn(txn ReadTxn, ws memdb.WatchSet, node string) ([]*structs.Allocation, error) {
 	// Get an iterator over the node allocations, using only the
 	// node prefix which ignores the terminal status
 	iter, err := txn.Get("allocs", "node_prefix", node)
@@ -3493,8 +3640,7 @@ func (s *StateStore) UpsertVaultAccessor(index uint64, accessors []*structs.Vaul
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeleteVaultAccessors is used to delete a set of Vault Accessors
@@ -3514,8 +3660,7 @@ func (s *StateStore) DeleteVaultAccessors(index uint64, accessors []*structs.Vau
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // VaultAccessor returns the given Vault accessor
@@ -3625,8 +3770,7 @@ func (s *StateStore) UpsertSITokenAccessors(index uint64, accessors []*structs.S
 		return errors.Wrap(err, "index update failed")
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeleteSITokenAccessors is used to delete a set of Service Identity token accessors.
@@ -3647,8 +3791,7 @@ func (s *StateStore) DeleteSITokenAccessors(index uint64, accessors []*structs.S
 		return errors.Wrap(err, "index update failed")
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // SITokenAccessor returns the given Service Identity token accessor.
@@ -3729,8 +3872,8 @@ func (s *StateStore) SITokenAccessorsByNode(ws memdb.WatchSet, nodeID string) ([
 
 // UpdateDeploymentStatus is used to make deployment status updates and
 // potentially make a evaluation
-func (s *StateStore) UpdateDeploymentStatus(index uint64, req *structs.DeploymentStatusUpdateRequest) error {
-	txn := s.db.WriteTxn(index)
+func (s *StateStore) UpdateDeploymentStatus(msgType structs.MessageType, index uint64, req *structs.DeploymentStatusUpdateRequest) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	if err := s.updateDeploymentStatusImpl(index, req.DeploymentUpdate, txn); err != nil {
@@ -3751,8 +3894,7 @@ func (s *StateStore) UpdateDeploymentStatus(index uint64, req *structs.Deploymen
 		}
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // updateDeploymentStatusImpl is used to make deployment status updates
@@ -3804,8 +3946,7 @@ func (s *StateStore) UpdateJobStability(index uint64, namespace, jobID string, j
 		return err
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // updateJobStabilityImpl updates the stability of the given job and version
@@ -3833,8 +3974,8 @@ func (s *StateStore) updateJobStabilityImpl(index uint64, namespace, jobID strin
 
 // UpdateDeploymentPromotion is used to promote canaries in a deployment and
 // potentially make a evaluation
-func (s *StateStore) UpdateDeploymentPromotion(index uint64, req *structs.ApplyDeploymentPromoteRequest) error {
-	txn := s.db.WriteTxn(index)
+func (s *StateStore) UpdateDeploymentPromotion(msgType structs.MessageType, index uint64, req *structs.ApplyDeploymentPromoteRequest) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	// Retrieve deployment and ensure it is not terminal and is active
@@ -3970,14 +4111,13 @@ func (s *StateStore) UpdateDeploymentPromotion(index uint64, req *structs.ApplyD
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // UpdateDeploymentAllocHealth is used to update the health of allocations as
 // part of the deployment and potentially make a evaluation
-func (s *StateStore) UpdateDeploymentAllocHealth(index uint64, req *structs.ApplyDeploymentAllocHealthRequest) error {
-	txn := s.db.WriteTxn(index)
+func (s *StateStore) UpdateDeploymentAllocHealth(msgType structs.MessageType, index uint64, req *structs.ApplyDeploymentAllocHealthRequest) error {
+	txn := s.db.WriteTxnMsgT(msgType, index)
 	defer txn.Abort()
 
 	// Retrieve deployment and ensure it is not terminal and is active
@@ -4066,8 +4206,7 @@ func (s *StateStore) UpdateDeploymentAllocHealth(index uint64, req *structs.Appl
 		}
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // LastIndex returns the greatest index value for all indexes
@@ -4278,8 +4417,7 @@ func (s *StateStore) ReconcileJobSummaries(index uint64) error {
 	if err := txn.Insert("index", &IndexEntry{"job_summary", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // setJobStatuses is a helper for calling setJobStatus on multiple jobs by ID.
@@ -4956,8 +5094,7 @@ func (s *StateStore) UpsertACLPolicies(index uint64, policies []*structs.ACLPoli
 		return fmt.Errorf("index update failed: %v", err)
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeleteACLPolicies deletes the policies with the given names
@@ -4974,8 +5111,7 @@ func (s *StateStore) DeleteACLPolicies(index uint64, names []string) error {
 	if err := txn.Insert("index", &IndexEntry{"acl_policy", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // ACLPolicyByName is used to lookup a policy by name
@@ -5063,8 +5199,7 @@ func (s *StateStore) UpsertACLTokens(index uint64, tokens []*structs.ACLToken) e
 	if err := txn.Insert("index", &IndexEntry{"acl_token", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // DeleteACLTokens deletes the tokens with the given accessor ids
@@ -5081,8 +5216,7 @@ func (s *StateStore) DeleteACLTokens(index uint64, ids []string) error {
 	if err := txn.Insert("index", &IndexEntry{"acl_token", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // ACLTokenByAccessorID is used to lookup a token by accessor ID
@@ -5216,8 +5350,7 @@ func (s *StateStore) BootstrapACLTokens(index, resetIndex uint64, token *structs
 	if err := txn.Insert("index", &IndexEntry{"acl_token_bootstrap", index}); err != nil {
 		return fmt.Errorf("index update failed: %v", err)
 	}
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // SchedulerConfig is used to get the current Scheduler configuration.
@@ -5246,8 +5379,7 @@ func (s *StateStore) SchedulerSetConfig(index uint64, config *structs.SchedulerC
 
 	s.schedulerSetConfigTxn(index, tx, config)
 
-	tx.Commit()
-	return nil
+	return tx.Commit()
 }
 
 func (s *StateStore) ClusterMetadata(ws memdb.WatchSet) (*structs.ClusterMetadata, error) {
@@ -5276,20 +5408,19 @@ func (s *StateStore) ClusterSetMetadata(index uint64, meta *structs.ClusterMetad
 		return errors.Wrap(err, "set cluster metadata failed")
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // WithWriteTransaction executes the passed function within a write transaction,
 // and returns its result.  If the invocation returns no error, the transaction
 // is committed; otherwise, it's aborted.
-func (s *StateStore) WithWriteTransaction(index uint64, fn func(Txn) error) error {
-	tx := s.db.WriteTxn(index)
+func (s *StateStore) WithWriteTransaction(msgType structs.MessageType, index uint64, fn func(Txn) error) error {
+	tx := s.db.WriteTxnMsgT(msgType, index)
 	defer tx.Abort()
 
 	err := fn(tx)
 	if err == nil {
-		tx.Commit()
+		return tx.Commit()
 	}
 	return err
 }
@@ -5317,7 +5448,9 @@ func (s *StateStore) SchedulerCASConfig(index, cidx uint64, config *structs.Sche
 
 	s.schedulerSetConfigTxn(index, tx, config)
 
-	tx.Commit()
+	if err := tx.Commit(); err != nil {
+		return false, err
+	}
 	return true, nil
 }
 
@@ -5374,8 +5507,7 @@ func (s *StateStore) UpsertScalingPolicies(index uint64, scalingPolicies []*stru
 		return err
 	}
 
-	txn.Commit()
-	return nil
+	return txn.Commit()
 }
 
 // upsertScalingPolicy is used to insert a new scaling policy.
@@ -5443,7 +5575,7 @@ func (s *StateStore) DeleteScalingPolicies(index uint64, ids []string) error {
 
 	err := s.DeleteScalingPoliciesTxn(index, ids, txn)
 	if err == nil {
-		txn.Commit()
+		return txn.Commit()
 	}
 
 	return err
@@ -5628,6 +5760,36 @@ func (s *StateStore) ScalingPoliciesByIDPrefix(ws memdb.WatchSet, namespace stri
 	return iter, nil
 }
 
+// LatestEventsReverse returns the unfiltered list of all volumes
+func (s *StateStore) LatestEventsReverse(ws memdb.WatchSet) (memdb.ResultIterator, error) {
+	txn := s.db.ReadTxn()
+	defer txn.Abort()
+
+	iter, err := txn.GetReverse("events", "id")
+	if err != nil {
+		return nil, fmt.Errorf("events lookup failed: %v", err)
+	}
+
+	ws.Add(iter.WatchCh())
+
+	return iter, nil
+}
+
+// Events returns the unfiltered list of all volumes
+func (s *StateStore) Events(ws memdb.WatchSet) (memdb.ResultIterator, error) {
+	txn := s.db.ReadTxn()
+	defer txn.Abort()
+
+	iter, err := txn.Get("events", "id")
+	if err != nil {
+		return nil, fmt.Errorf("events lookup failed: %v", err)
+	}
+
+	ws.Add(iter.WatchCh())
+
+	return iter, nil
+}
+
 // scalingPolicyNamespaceFilter returns a filter function that filters all
 // scaling policies not targeting the given namespace.
 func scalingPolicyNamespaceFilter(namespace string) func(interface{}) bool {
@@ -5639,6 +5801,18 @@ func scalingPolicyNamespaceFilter(namespace string) func(interface{}) bool {
 
 		return p.Target[structs.ScalingTargetNamespace] != namespace
 	}
+}
+
+// UpsertEvents is used to insert events. It should only be used for testing.
+// Normal use events are inserted to go-memdb during transaction commit
+func (s *StateStore) UpsertEvents(index uint64, events *structs.Events) error {
+	txn := s.db.WriteTxn(index)
+	defer txn.Abort()
+
+	if err := txn.Insert("events", events); err != nil {
+		return err
+	}
+	return txn.Commit()
 }
 
 // StateSnapshot is used to provide a point-in-time snapshot
@@ -5742,8 +5916,8 @@ func (s *StateRestore) Abort() {
 }
 
 // Commit is used to commit the restore operation
-func (s *StateRestore) Commit() {
-	s.txn.Commit()
+func (s *StateRestore) Commit() error {
+	return s.txn.Commit()
 }
 
 // NodeRestore is used to restore a node
@@ -5884,6 +6058,13 @@ func (r *StateRestore) CSIPluginRestore(plugin *structs.CSIPlugin) error {
 func (r *StateRestore) CSIVolumeRestore(volume *structs.CSIVolume) error {
 	if err := r.txn.Insert("csi_volumes", volume); err != nil {
 		return fmt.Errorf("csi volume insert failed: %v", err)
+	}
+	return nil
+}
+
+func (r *StateRestore) EventRestore(events *structs.Events) error {
+	if err := r.txn.Insert("events", events); err != nil {
+		return fmt.Errorf("events insert failed: %v", err)
 	}
 	return nil
 }